+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 - fix: minor UI and logics fix in the service-worker
 - style: Add a max width to the bottom menu
 - fix: minor errors in the bottom menu icons
+- fix: minor errors in the reading top banner save button
 
 ## [3.4.0]